feat: move resources fns to utils + add tests
